+++ src/main/java/com/epam/esm/repository/api/TagGiftCertificateRepository.java
@@ -14,7 +14,7 @@ public interface TagGiftCertificateRepository {
      * @param tagId         The ID of the Tag entity to associate with the GiftCertificate entity.
      * @return The number of rows affected by the creation operation (should always be 1 in this case).
      */
-    int createTagGiftCertificate(int certificateId, int tagId);
+    int createTagGiftCertificate(long certificateId, long tagId);
 
     /**
      * Retrieves a list of Tag entity IDs that are associated with a GiftCertificate entity with the specified ID in the repository.
@@ -22,7 +22,7 @@ public interface TagGiftCertificateRepository {
      * @param certificateId The ID of the GiftCertificate entity to retrieve associated Tag entity IDs for.
      * @return A list of Tag entity IDs that are associated with the specified GiftCertificate entity.
      */
-    List<Integer> getAllTagsIdByGiftCertificate(int certificateId);
+    List<Long> getAllTagsIdByGiftCertificate(long certificateId);
 
     /**
      * Retrieves a list of GiftCertificate entity IDs that are associated with a Tag entity with the specified ID in the repository.
@@ -30,5 +30,5 @@ public interface TagGiftCertificateRepository {
      * @param tagId The ID of the Tag entity to retrieve associated GiftCertificate entity IDs for.
      * @return A list of GiftCertificate entity IDs that are associated with the specified Tag entity.
      */
-    List<Integer> getAllCertificateIdByTag(int tagId);
+    List<Long> getAllCertificateIdByTag(long tagId);
 }
